Moved sidebar out of microhook
Moved the sidebar out of the after-post-list microhook and into the base theme template. This should resolve the styling issue which will break sites using the before-post-list microhook. This does tie the plugin to Tiny-theme baseof theme, which is not great. Will need to resolve this in some way, but should work for now.